+++ Src/Melville.Pdf.LowLevel/Filters/Jbig2Filter/BinaryBitmaps/OffsetReader.cs
@@ -3,9 +3,9 @@
 public ref struct OffsetReader
 {
     private readonly byte readOffset;
-    private ushort buffer;
+    private int buffer;
 
-    public OffsetReader(byte offset, ushort buffer = 0 ) : this()
+    public OffsetReader(byte offset, int buffer = 0 ) : this()
     {
         readOffset = (byte)((offset == 0)?0:(8 - offset));
         this.buffer = buffer;
@@ -15,7 +15,7 @@ public ref struct OffsetReader
     {
         //rejected an alternate explanation where we let readoffser = 8 is rejected because that
         // would read one extra byte off the end of the array.
-        if (readOffset > 0) buffer = (ushort)(*src++);
+        if (readOffset > 0) buffer = (int)(*src++);
     }
 
     public unsafe byte ReadBye(ref byte* src)
